Park func_800336F0 at 2, and func_8005AE68 26 -> 22
func_800336F0 was two step-0 reads away from close: gp == 0 with three $at
stores means recipe branch 1 -- real scalar declarations plus `as -G0`, with
D_8009B254 needing its _IS_SCALAR arm because its default is an unsized array
and that gives cc1psx's own pair rather than the assembler's $at. The other
was D_8009B398_IS_VOLATILE: its two tests are on different paths, so two
loads look automatic, but gcc's CSE reaches across the branch and commons
them while retail reloads. -3 to 0 on that word, and 78 differences to 2 with
the `addu base,index` spelling.

What is left is one constant materialised a branch too early. Five shapes
measured against it, all identical.

func_8005AE68 improved to 22 from the permuter's second run, and this one is
a real find rather than an artifact: `s16 x` instead of `s32` for the
packed-colour temp -- the same narrow-type lever as the `u8 m` that was
already the key to that function. Narrowing y, lim or f the same way is +2,
-1 and +3, so it is that one variable. Its residue is now nothing but the
$s3/$s4/$s5 rotation.

Build green, sha1 verified.

+++ parked/func_8005AE68.c
@@ -25,7 +25,10 @@ s32 func_8005AE68(s32 arg0, s32 arg1, s32 arg2) {
     s32 h;
     u8 m;
     s32 e6;
-    s32 x;
+    /* s16, not s32: the permuter found it and it is the same lever as the
+     * `u8 m` above -- the narrower type is what makes gcc keep the
+     * truncation. 26 differences to 22. */
+    s16 x;
     s32 y;
     s32 a;
     s32 b;
